+++ CHANGELOG.md
@@ -2,11 +2,15 @@
 
 ## 1.0.2
 
+IMPROVEMENT
+
+- [config] Add new seed nodes
+- [api] Add /missed_blocks endpoint
+
 BUG FIXES
 
 - [api] Fix block proposer issue
 - [api] Fix 0x api error
-- [api] Add /missed_blocks endpoint
 
 ## 1.0.0